Trying to properly register new sensors in sensor.py and log more details when culling cached commands.
